chore: reorganized components into feature specific directories throughout the app
