+++ src/components/dashboard/Dashboard.js
@@ -6,7 +6,7 @@ import { compose } from 'redux';
 import { firestoreConnect } from 'react-redux-firebase';
 import PropTypes from 'prop-types';
 
-import ProjectList from '../projects/ProjectList';
+import Projects from '../projects/Projects';
 import Notifications from './Notifications';
 import Spaces from './Spaces';
 import { getMeetupsThunkCreator } from '../../store/reducers/meetupsReducer';
@@ -32,7 +32,7 @@ const Dashboard = ({
       <div className="dashboard container">
         <div className="row">
           <div className="col s12 m6">
-            <ProjectList projects={projects} />
+            <Projects projects={projects} />
           </div>
 
           <div className="col s12 m5 offset-m1">
